+++ src/PHPCR/Shell/Console/Command/Phpcr/LockUnlockCommand.php
@@ -30,15 +30,6 @@ Also removes the properties jcr:lockOwner and jcr:lockIsDeep from that
 node. As well, the corresponding lock token is removed from the set of
 lock tokens held by the current Session.
 
-If the node does not currently hold a lock or holds a lock for which
-this Session is not the owner and is not a "lock-superuser", then a
-\PHPCR\Lock\LockException is thrown.
-
-<b>Note:</b>
-However that the system may give permission to a non-owning session
-to unlock a lock. Typically such "lock-superuser" capability is intended
-to facilitate administrational clean-up of orphaned open-scoped locks.
-
 Note that it is possible to unlock a node even if it is checked-in (the
 lock-related properties will be changed despite the checked-in status).
 HERE