+++ src/PHPCR/Shell/Console/Command/Phpcr/NodeMixinAddCommand.php
@@ -30,21 +30,6 @@ If this node is already of type <info>mixinName</info> (either due to a previous
 added mixin or due to its primary type, through inheritance) then this
 method has no effect. Otherwise <info>mixinName</info> is added to this node's
 jcr:mixinTypes property.
-
-Semantically, the new node type may take effect immediately, on dispatch
-or on persist. The behavior is adopted must be the same as the behavior
-adopted for NodeInterface::setPrimaryType() and the behavior that
-occurs when a node is first created.
-
-A ConstraintViolationException is thrown either immediately or on save
-if a conflict with another assigned mixin or the primary node type
-occurs or for an implementation-specific reason. Implementations may
-differ on when this validation is done.
-
-In some implementations it may only be possible to add mixin types
-before a a node is persisted for the first time. In such cases any
-later calls to <info>addMixin</info> will throw a ConstraintViolationException
-either immediately, on dispatch or on persist.
 HERE
         );
     }