+++ src/PHPCR/Shell/Console/Command/Phpcr/SessionImpersonateCommand.php
@@ -29,19 +29,6 @@ Note: This command is not implemented by any of the transports currently.
 
 Returns a new session in accordance with the specified (new)
 Credentials.
-
-Allows the current user to acquire a new session using incomplete or
-relaxed credentials requirements (perhaps including a user name but no
-password, for example), assuming that this Session gives them that
-permission. This method can be used to "impersonate" another user or to
-clone the current session by passing in the same credentials that were
-used to acquire the current session.
-
-The new Session is tied to a new Workspace instance. In other words,
-Workspace instances are not re-used. However, the Workspace instance
-returned represents the same actual persistent workspace entity in the
-repository as is represented by the Workspace object tied to the current
-Session.
 HERE
         );