+++ src/PHPCR/Shell/Console/Command/Phpcr/SessionLoginCommand.php
@@ -25,6 +25,7 @@ class SessionLoginCommand extends BasePhpcrCommand
         $this->addArgument('password', InputArgument::REQUIRED, 'Password');
         $this->addArgument('workspaceName', InputArgument::OPTIONAL, 'Optional workspace name');
         $this->setHelp(<<<HERE
+Login to a session.
 HERE
         );
     }